[BREAKING CHANGE] Optional fields aren't Nullable!() now
This was redundant and annoying. Because `@optional` fields are equal to
`typeof(field).init` by default, code can be easily simplified.

This way, the following code used before this commit:
```
if(!message.text.isNull)
	writeln(message.text);
```

Is equal to this after this commit:
```
if(message.text == typeof(message.text).init)
	writeln(message.text);
```

Or, because we know that `message.text` is a `string`
(and we always do, because if it isn't a string, the following code won't compile)
we can use the following:
```
if(message.text.length)
	writeln(message.text)
```

In this example `.length` could be omitted too.

This is a breaking change, but since no release is tagged yet and updating existing bots
to this commit will be relatively easy, it will be commited right into master.

Also, this saves ~500MB of memory during compilation.